Create CourseValidator class for better OOP
