+++ src/main/java/seedu/exchangecoursemapper/command/AddCoursesCommand.java
@@ -5,88 +5,21 @@ import seedu.exchangecoursemapper.constants.Logs;
 import seedu.exchangecoursemapper.courses.Course;
 import seedu.exchangecoursemapper.exception.Exception;
 import seedu.exchangecoursemapper.storage.Storage;
+import seedu.exchangecoursemapper.parser.CourseValidator;
 
-import javax.json.JsonArray;
 import javax.json.JsonObject;
 import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import static seedu.exchangecoursemapper.constants.JsonKey.*;
-import static seedu.exchangecoursemapper.constants.Messages.LINE_SEPARATOR;
-import static seedu.exchangecoursemapper.constants.Messages.LIST_RELEVANT_PU;
-
 
 public class AddCoursesCommand extends PersonalTrackerCommand {
 
-    public static final String RETRIEVE_COURSE_LIST = "Retrieve course list for partner university";
     private static final Logger logger = Logger.getLogger(AddCoursesCommand.class.getName());
-    public static final String EMPTY_COURSE_INPUT = "NUS or Partner University course input cannot be null.";
-    public static final String NO_COURSES_AVAILABLE = "No courses available for the partner university: ";
+    private CourseValidator courseValidator;
 
-    private static boolean isValidCourseMapping(String nusCourseInput, String puCourseInput,
-                                                JsonArray courses, String pu) {
-
-        if (nusCourseInput.isEmpty() || puCourseInput.isEmpty()) {
-            throw new IllegalArgumentException(EMPTY_COURSE_INPUT);
-        }
-        if (courses == null || courses.isEmpty()) {
-            throw new IllegalArgumentException(NO_COURSES_AVAILABLE + pu);
-        }
-
-        for (int i = 0; i < courses.size(); i++) {
-            logger.log(Level.INFO, Logs.FIND_COURSE_MAPPING);
-            JsonObject course = courses.getJsonObject(i);
-            String puCourseCode = course.getString(PU_COURSE_CODE_KEY).toLowerCase();
-            String nusCourseCode = course.getString(NUS_COURSE_CODE_KEY).toLowerCase();
-
-            if (puCourseCode.equals(puCourseInput) && nusCourseCode.equals(nusCourseInput)) {
-                return true;
-            }
-        }
-
-        System.out.println("Invalid course mapping!");
-        displayAvailableMappings(courses, pu);
-        return false;
-    }
-
-    private static void displayAvailableMappings(JsonArray courses, String pu) {
-        System.out.println("The available mappings for " + pu + " are:");
-        System.out.println(LINE_SEPARATOR);
-
-        for (int i = 0; i < courses.size(); i++) {
-            JsonObject course = courses.getJsonObject(i);
-            String puCourseCode = course.getString(PU_COURSE_CODE_KEY).toLowerCase();
-            String nusCourseCode = course.getString(NUS_COURSE_CODE_KEY).toLowerCase();
-            String nusCourseName = course.getString(NUS_COURSE_NAME_KEY).toLowerCase();
-            String puCourseName = course.getString(PU_COURSE_NAME_KEY).toLowerCase();
-
-            System.out.println(nusCourseCode + " " + nusCourseName + " | " + puCourseCode + " " + puCourseName + System.lineSeparator());
-        }
-        System.out.println(LINE_SEPARATOR);
-    }
-
-    private static JsonArray getPUCourseList(String pu, JsonObject jsonObject) {
-        JsonArray courses;
-        logger.log(Level.INFO, Logs.FIND_PARTNER_UNIVERSITY);
-        String matchPu = jsonObject.keySet().stream().filter(key -> key.equalsIgnoreCase(pu)).findFirst().orElse(null);
-
-        logger.log(Level.INFO, Logs.UNIVERSITY_FOUND);
-        if (matchPu != null) {
-            logger.log(Level.INFO, RETRIEVE_COURSE_LIST);
-            courses = jsonObject.getJsonObject(matchPu).getJsonArray(COURSES_ARRAY_LABEL);
-        } else {
-            logger.log(Level.INFO, Logs.INVALID_UNIVERSITY_INPUT);
-            System.out.println(Logs.INVALID_UNIVERSITY_INPUT);
-
-            logger.log(Level.INFO, Logs.DISPLAY_PARTNER_UNIVERSITIES);
-            System.out.println(LINE_SEPARATOR);
-            System.out.println(LIST_RELEVANT_PU);
-            System.out.println(LINE_SEPARATOR);
-
-            throw new IllegalArgumentException(Logs.INVALID_UNIVERSITY_INPUT);
-        }
-        return courses;
+    public AddCoursesCommand() {
+        courseValidator = new CourseValidator();
     }
 
     @Override
@@ -109,7 +42,7 @@ public class AddCoursesCommand extends PersonalTrackerCommand {
             String puCourse = descriptionSubstrings[2].trim().toLowerCase();
 
             logger.log(Level.INFO, Logs.FORMAT);
-            boolean isValidInput = isValidInput(nusCourse, pu, puCourse, jsonObject);
+            boolean isValidInput = courseValidator.isValidInput(nusCourse, pu, puCourse, jsonObject);
 
             if (isValidInput) {
                 logger.log(Level.INFO, Logs.ADD_APPROVED_MAPPING);
@@ -169,22 +102,4 @@ public class AddCoursesCommand extends PersonalTrackerCommand {
         System.out.println("You have successfully added the course: " + addCourse.formatOutput());
     }
 
-    public boolean isValidInput(String nusCourseInput, String pu, String puCourseInput, JsonObject jsonObject) {
-
-        assert nusCourseInput != null : "NUS course should not be empty";
-        assert pu != null : "Partner university should not be empty";
-        assert puCourseInput != null : "Partner university course should not be empty";
-        assert jsonObject != null : "JSON object should not be empty";
-
-        logger.log(Level.INFO, Logs.CHECK_UNIVERSITY);
-        JsonArray courses = getPUCourseList(pu, jsonObject);
-
-        if (courses == null) {
-            logger.log(Level.WARNING, Logs.INVALID_UNIVERSITY_INPUT);
-            throw new IllegalArgumentException(Logs.INVALID_UNIVERSITY_INPUT);
-        }
-
-        logger.log(Level.INFO, Logs.CHECK_COURSE_MAPPING);
-        return isValidCourseMapping(nusCourseInput, puCourseInput, courses, pu);
-    }
 }
